Add support for GnuPG 2.1 ('modern') while retaining the original behaviour with GnuPG 2.0 ('classic'), and for running the script under MacPorts
Compared to v2.0, v2.1:

 - requires '--full-gen-key' rather than '--gen-key' to give access to the full key generation options;
 - automatically generates a revocation certificate when generating a master key pair;
 - includes keygrip information in key listings;
 - simplifies the removal of a private key by turning it into a single file deletion operation (with the filename being based on the keygrip).

The script now parses the gpg2 version string to determine whether v2.0 or v2.1 is being used, sets the variable EGK_GPGMODERN to 'false' or 'true' and the variable EGK_GPGGENKEY to '--gen-key' or '--full-gen-key' respectively and uses the value of EGK_GPGMODERN to control the way various operations that differ between v2.0 and v2.1 are performed. Keygrip information (present with v2.1) is parsed from the colon-delimited key listings and included in the YAML output and used to identify the master private key file (if the master private key is to be removed). Revocation certificate generation is skipped with v2.1 since one is automatically created along with the master key pair. Securly deleting files accounts for MacPorts installing GNU 'shred' as 'gshred'.